コメントの追加＋メニューのプライオリティを０に設定

+++ player.x.h
@@ -65,7 +65,7 @@ private: //アクセス不可能
 	constexpr static int   MAX_TEXT_OBJECT = 1;             //今配置ツールに設定されてるオブジェクトの数ー１（０スタートだから）
 	constexpr static int   MAX_BUULET_SPEED = 30;           //弾の速さ
 	constexpr static int   MAX_DIE_COUNT = 180;             //死亡時のカウント
-	constexpr static int   ADDJUST_SELECT_SIZEY = 160;      //選択ゲージのY軸の位置を加算する値
+	constexpr static float   ADDJUST_SELECT_SIZEY = 160.0f;      //選択ゲージのY軸の位置を加算する値
 
 	//ぶっ飛び処理時マクロ定義
 	constexpr static int   BLOWAWAY_FRAME_1 = 40;           //第一行動時のフレーム
